Fix firefox drag bug, click & drag to create event, color picker
Unfortunately there is a fixed set of colors that can be used for
events with google. Regardless, a color can be selected for an event.

Firefox requires data to be assigned to a drag event in order for it to
carry through.

Clicking on an empty space somewhere on a Day resulted in an hour-long
event being created there. If the user clicks and drags on an empty
space, they can create a variable-length event.